Move location of Chain constants to new eth.constants module
