+++ python/examples/resolve_eth_names.py
@@ -3,7 +3,7 @@ Resolves ENS domains to Ethereum addresses
 ENS is a smart-contract system that registers and resolves `.eth` domains.
 """
 import in3
-from in3.model import Chain
+from in3.eth.constants import Chain
 
 
 def _print():
