+++ python/examples/send_transaction.py
@@ -5,7 +5,7 @@ Works with included `data` field for smart-contract calls.
 """
 import json
 import in3
-from in3.model import Chain
+from in3.eth.constants import Chain
 import time
