+++ python/in3/eth/constants.py
@@ -1,0 +1,7 @@
+class Chain:
+    MAINNET = "mainnet"
+    KOVAN = "kovan"
+    EVAN = "evan"
+    GOERLI = "goerli"
+    IPFS = "ipfs"
+    EWC = "ewc"
