+++ python/tests/integrated/negative/client_test.py
@@ -4,7 +4,7 @@ Integrated negative tests for `in3` module. Doesnt test submodules.
 import unittest
 
 import in3
-from in3.model import Chain
+from in3.eth.constants import Chain
 from tests.integrated.mock.config import mock_config
 from tests.integrated.mock.transport import mock_transport
